update kill logic

+++ src/main/java/game/user/GameProfile.java
@@ -12,8 +12,8 @@ import java.util.Random;
  */
 public class GameProfile {
     private int score;
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private double direction;
     private Instant dt;
     private boolean isKilled;
@@ -43,19 +43,19 @@ public class GameProfile {
         this.score = score;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
